+++ workers/unity/Assets/Playground/Scripts/Worker/iOSClientWorkerConnector.cs
@@ -16,9 +16,11 @@ namespace Playground
         [SerializeField] private GameObject level;
 
         private GameObject levelInstance;
+        private ConnectionService connectionService;
 
-        public async void TryConnect()
+        public async void TryConnectAsync(ConnectionService connectionService)
         {
+            this.connectionService = connectionService;
             await Connect(WorkerUtils.iOSClient, new ForwardingDispatcher()).ConfigureAwait(false);
         }
 
@@ -53,6 +55,11 @@ namespace Playground
             throw new PlatformNotSupportedException(
                 $"{nameof(iOSClientWorkerConnector)} can only be used for the iOS platform. Please check your build settings.");
 #endif
+        }
+
+        protected override ConnectionService GetConnectionService()
+        {
+            return connectionService;
         }
 
         public override void Dispose()