+++ workers/unity/Packages/com.improbable.gdk.core/Config/RuntimeConfig.cs
@@ -12,6 +12,7 @@ namespace Improbable.Gdk.Core
         public const string LocatorHost = "locator.improbable.io";
         public const string ReceptionistHost = "127.0.0.1";
         public const ushort ReceptionistPort = 7777;
+        public const ushort AnonymousAuthenticationPort = 444;
     }
 
     /// <summary>