+++ workers/unity/Packages/com.improbable.gdk.mobile/Worker/MobileWorkerConnector.cs
@@ -1,10 +1,15 @@
+using System;
 using Improbable.Gdk.Core;
 using Improbable.Worker.CInterop;
+using Improbable.Worker.CInterop.Alpha;
+using UnityEngine;
 
 namespace Improbable.Gdk.Mobile
 {
     public abstract class MobileWorkerConnector : WorkerConnector
     {
+        [SerializeField] private string DevelopmentAuthToken;
+        
         protected abstract string GetHostIp();
 
         protected override ConnectionParameters GetConnectionParameters(string workerType, ConnectionService service)
@@ -34,17 +39,28 @@ namespace Improbable.Gdk.Mobile
 
         protected override LocatorConfig GetLocatorConfig()
         {
-            throw new System.NotImplementedException("The locator flow is currently not available for mobile workers.");
+            throw new NotImplementedException("The locator flow is currently not available for mobile workers.");
         }
 
-        protected override AlphaLocatorConfig GetAlphaLocatorConfig()
+        protected override AlphaLocatorConfig GetAlphaLocatorConfig(string workerType)
         {
-            throw new System.NotImplementedException("The alpha locator flow is currently not available for mobile workers.");
-        }
+            var pit = GetDevelopmentPlayerIdentityToken(DevelopmentAuthToken, GetPlayerId(), GetDisplayName());
+            var loginTokenDetails = GetDevelopmentLoginTokens(workerType, pit);
+            var loginToken = SelectLoginToken(loginTokenDetails);
 
-        protected override ConnectionService GetConnectionService()
-        {
-            return ConnectionService.Receptionist;
+            return new AlphaLocatorConfig
+            {
+                LocatorHost = RuntimeConfigDefaults.LocatorHost,
+                LocatorParameters = new Worker.CInterop.Alpha.LocatorParameters
+                {
+                    PlayerIdentity = new PlayerIdentityCredentials
+                    {
+                        PlayerIdentityToken  = pit,
+                        LoginToken = loginToken,
+                    },
+                    UseInsecureConnection = false,
+                }
+            };
         }
     }
 }